feat(abstract-utxo): rename signPsbt to signPsbtUtxolib
Rename signPsbt.ts to signPsbtUtxolib.ts and update all references to
it. This improves clarity by making the utxolib dependency explicit in
the filename. Also rename signPsbtWithMusig2Participant to
signPsbtWithMusig2ParticipantUtxolib for consistency.

Issue: BTC-2909
